refactor: deepen integrity panel
